[Rework][Back] Moving the class Author into the data/internal package
